Reloaded : Add Kf Anim toggle button in curve panel
When key framing an fx parameter, it's a bad user experience to have
to (open and) go to the fx chain panel to active the animation.

I had chosen to duplicate the toogle button to the curve (in place of just move it from
fx chain), because it's relevant to me to have this functionnaly near the fx chain.
(i'm also thinking to change the chain toogle by modifier+click on chain bulbs someday for both
fx activation state and fx kf anim status)